Remove extraneous function
