+++ packages/core/src/utils.ts
@@ -973,25 +973,6 @@ export function splitN(str: string, delim: string, n: number): string[] {
 }
 
 /**
- * Splits a string on the last occurrence of the delimiter
- * @param str - The string to split
- * @param delim - The delimiter string
- * @returns An array of two strings; the first consisting of the beginning of the
- * string up to the last occurrence of the delimiter. the second is the remainder of the
- * string after the last occurrence of the delimiter. If the delimiter is not present
- * in the string, the first element is empty and the second is the input string.
- */
-export function splitOnceRight(str: string, delim: string): [string, string] {
-  const delimIndex = str.lastIndexOf(delim);
-  if (delimIndex === -1) {
-    return ['', str];
-  }
-  const beginning = str.substring(0, delimIndex);
-  const last = str.substring(delimIndex + delim.length);
-  return [beginning, last];
-}
-
-/**
  * Memoizes the result of a parameterless function
  * @param fn - The function to be wrapped
  * @returns The result of the first invocation of the wrapped function
